docs: rename mini_assembly to assembly_engine for clarity and consistency
