+++ tests/unit/internal/test_assembly_engine.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import pytest
 
-from diwire._internal.resolvers.assembly.mini_assembly import Environment
+from diwire._internal.resolvers.assembly.assembly_engine import Environment
 
 
 def test_variable_interpolation_renders_context_value() -> None:
